# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plexe ML Model Service",
    "description": "ML Model Management and Prediction Service",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/models/preview": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Preview Model",
        "description": "Preview model metadata before upload.",
        "operationId": "preview_model_api_v1_models_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_model_api_v1_models_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Preview Model Api V1 Models Preview Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/upload": {
      "post": {
        "tags": [
          "models"
        ],
        "summary": "Upload Model",
        "description": "Upload and deploy an XGBoost model with versioning support.",
        "operationId": "upload_model_api_v1_models_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_model_api_v1_models_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "List Models",
        "description": "Get all uploaded models or only latest versions.",
        "operationId": "list_models_api_v1_models_get",
        "parameters": [
          {
            "name": "latest_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Latest Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}": {
      "get": {
        "tags": [
          "models"
        ],
        "summary": "Get Model",
        "description": "Get specific model by ID.",
        "operationId": "get_model_api_v1_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelMeta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "models"
        ],
        "summary": "Delete Model",
        "description": "Delete a model.",
        "operationId": "delete_model_api_v1_models__model_id__delete",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}/predict": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Predict",
        "description": "Make a prediction using the specified model.",
        "operationId": "predict_api_v1_models__model_id__predict_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/models/{model_id}/predict/batch": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Predict Batch",
        "description": "Make batch predictions using the specified model.",
        "operationId": "predict_batch_api_v1_models__model_id__predict_batch_post",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ws/status": {
      "get": {
        "tags": [
          "websocket"
        ],
        "summary": "Websocket Status",
        "description": "Get WebSocket connection status.",
        "operationId": "websocket_status_api_v1_ws_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Root",
        "description": "Root endpoint with service information.",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "Health check endpoint.",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_preview_model_api_v1_models_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_model_api_v1_models_preview_post"
      },
      "Body_upload_model_api_v1_models_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_new_version": {
            "type": "boolean",
            "title": "Is New Version",
            "default": false
          },
          "parent_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Model Id"
          }
        },
        "type": "object",
        "required": [
          "file",
          "name"
        ],
        "title": "Body_upload_model_api_v1_models_upload_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ModelListResponse": {
        "properties": {
          "models": {
            "items": {
              "$ref": "#/components/schemas/ModelMeta"
            },
            "type": "array",
            "title": "Models",
            "description": "List of models"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of models"
          }
        },
        "type": "object",
        "required": [
          "models",
          "total"
        ],
        "title": "ModelListResponse"
      },
      "ModelMeta": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Model name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Model description"
          },
          "model_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of model"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique model identifier"
          },
          "status": {
            "$ref": "#/components/schemas/ModelStatus",
            "default": "uploaded"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path",
            "description": "Path to model file"
          },
          "feature_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Names",
            "description": "Expected input features"
          },
          "model_info": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Info",
            "description": "Additional model information"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "Model version number",
            "default": 1
          },
          "parent_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Model Id",
            "description": "ID of the parent model for versioning"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest",
            "description": "Whether this is the latest version of the model",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "ModelMeta"
      },
      "ModelStatus": {
        "type": "string",
        "enum": [
          "uploaded",
          "deployed",
          "error"
        ],
        "title": "ModelStatus"
      },
      "ModelType": {
        "type": "string",
        "enum": [
          "classifier",
          "regressor"
        ],
        "title": "ModelType"
      },
      "PredictionBatchRequest": {
        "properties": {
          "features": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Features",
            "description": "Batch of input features"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "PredictionBatchRequest"
      },
      "PredictionBatchResponse": {
        "properties": {
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/PredictionResponse"
            },
            "type": "array",
            "title": "Predictions",
            "description": "Batch predictions"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "description": "ID of model used for predictions"
          }
        },
        "type": "object",
        "required": [
          "predictions",
          "model_id"
        ],
        "title": "PredictionBatchResponse"
      },
      "PredictionRequest": {
        "properties": {
          "features": {
            "type": "object",
            "title": "Features",
            "description": "Input features for prediction"
          }
        },
        "type": "object",
        "required": [
          "features"
        ],
        "title": "PredictionRequest"
      },
      "PredictionResponse": {
        "properties": {
          "prediction": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "integer"
              },
              {
                "type": "string"
              }
            ],
            "title": "Prediction",
            "description": "Model prediction"
          },
          "probability": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Probability",
            "description": "Prediction probability (for classifiers)"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence",
            "description": "Confidence level description"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id",
            "description": "ID of model used for prediction"
          }
        },
        "type": "object",
        "required": [
          "prediction",
          "model_id"
        ],
        "title": "PredictionResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}